Fix bug
In the ovelap solveur, when trying to make sure all filled intervals are in a possible interval there was a bug. While selecting which interval should have the filled interval inside, we need to take the one whose *min* value is before the interval and not the *max* value because the filled interval can already be partly in it.